requirements.txt
langgraph>=0.0.15
langchain>=0.1.0
langchain-core>=0.1.17
langchain-groq>=0.0.3
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.118.0
python-dotenv>=1.0.0
rich>=13.7.0
PyYAML>=6.0.1
langchain_community